Fix/refactoring

+++ app/src/main/kotlin/com/evastos/bux/data/model/rtf/update/UpdateEventBody.kt
@@ -2,9 +2,10 @@ package com.evastos.bux.data.model.rtf.update
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
+import java.math.BigDecimal
 
 @JsonClass(generateAdapter = true)
 data class UpdateEventBody(
     @Json(name = "securityId") val securityId: String?,
-    @Json(name = "currentPrice") val currentPrice: Double?
+    @Json(name = "currentPrice") val currentPrice: BigDecimal?
 )
